+++ src/js/directives/index.js
@@ -2,7 +2,7 @@
 
 module.exports = function(app) {
   require('./compareTo')(app);
-  require('./ipv4Address')(app);
+  require('./ipAddress')(app);
   require('./fileRequired')(app);
   require('./objectLength')(app);
 };
